+++ NEWS.md
@@ -27,6 +27,7 @@
   * Default savegame name now includes the WAD filename (thanks Fabian).
   * Mouse movement is no longer read when the game window is inactive
     (thanks Julia Nechaevskaya).
+  * Add native support for the FluidSynth midi synthesizer.
 
 ### Refactorings
   * CMake project files have been added, replacing the Microsoft Visual